Move gaussian.
