+++ mmdet3d/core/utils/__init__.py
@@ -1,0 +1,3 @@
+from .gaussian import draw_heatmap_gaussian, gaussian_2d, gaussian_radius
+
+__all__ = ['gaussian_2d', 'gaussian_radius', 'draw_heatmap_gaussian']
